perf: 1 サンプルごとの引き直しと、実体ごとの重複した表を減らす
・音源ごとの目標レートを毎サンプル switch で引き直していたのをやめ、
  レートが変わったときだけ求めて持っておくようにする
  (OPL / OPL3 / OPM / OPN / OPNA / OPZX7。WT2 / WT+ は
   すでに持っていた値を見ずに引き直していたので、そちらを見るだけ)
・音源の対応表 std::map<OscMode, SynthCore*> を素の配列にする。
  モード番号で引くだけなので木をたどる必要が無い
・キースケールの表 (KSOpl / KSMa7 / KSOpn / KSOpp / KSOps / KSOpz) を
  実体ごとに持つのをやめ、1 組だけ作って皆で見るようにする。
  中身は音符から決まるだけでどの実体でも同じ。オペレータ 1 本あたり
  最大 6KB あり、OPZX7 は 8 本 × ボイス数ぶん抱えていた

あわせて直したもの:
・KSOps の表の末尾 (音程差 +127) が埋まっておらず、最高音を
  ブレークポイント 0 で鳴らすと未初期化の値を読んでいた
・キースケールの各クラスで、setParameters より前に呼ぶと
  未初期化のメンバを読む状態だったので初期値を入れた

死にコードの掃除:
・どこからも読み込んでいない RateLUT.h を消す
・FM 各音源の m_history2 (毎サンプル配列を丸ごと写していたが
  誰も読まない) を消す
・使われていないレジスタ変換 9 個を消す

+++ 2686VLight/Source/Core/Processor/PluginProcessor.h
@@ -102,7 +102,7 @@ public:
                     // これにより、波形が強制キルされず、位相や音量が完全に引き継がれます。
                     if (voice->isVoiceActive()) {
                         auto cyclesPerSecond = juce::MidiMessage::getMidiNoteInHertz(midiNoteNumber);
-                        voice->coreMap[currentParams->mode]->noteOn(cyclesPerSecond, velocity, midiNoteNumber, isLegato);
+                        voice->coreMap[(size_t)currentParams->mode]->noteOn(cyclesPerSecond, velocity, midiNoteNumber, isLegato);
                     }
                     else {
                         // 完全に音が消えている時だけ、通常の startVoice でボイスを起こす
